Added Abbreviations for all economy based commands and formated numbers with commas.
Added Gamemode Shortcut Commands
Added Invesee Command
Added PacketRefresh Command
Added Spawn System, Setspawn command, and Spawn command
Added Tab Completion of Econ commands
Reworked Ranks to allow staff and normal ranks per player

+++ src/main/java/org/evasive/me/minefinity/admin/commands/economy/sub/EconomySetSub.java
@@ -8,6 +8,8 @@ import org.evasive.me.minefinity.core.SubCommand;
 import org.evasive.me.minefinity.economy.EconomyService;
 import org.evasive.me.minefinity.utils.CommandFeedback;
 
+import static org.evasive.me.minefinity.utils.economy.EconNumberUtils.*;
+
 public class EconomySetSub implements SubCommand {
 
     EconomyService economyService = Minefinity.getCore().getEconomyService();
@@ -22,7 +24,8 @@ public class EconomySetSub implements SubCommand {
 
         String playerName = args[1];
         Player target = Bukkit.getPlayerExact(playerName);
-        int amount = Integer.parseInt(args[2]);
+
+        double amount = Double.parseDouble(args[2]);
 
         economyService.setBalance(target, amount);
